Reorganise file structure, finish rewrite of parfit-login db script, add db action scripts build command
