+++ db/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "compilerOptions": {
     "isolatedModules": false,
-    "outDir": "../dist/rules",
+    "outDir": "../dist/db",
     "rootDir": "./",
     "target": "ES2020"
   },
