+++ rules/types/globals.ts
@@ -1,0 +1,5 @@
+/** Allow TEMPLATE_DATA to be used as a global in rule definitions */
+declare global {
+  const TEMPLATE_DATA: Record<string, any>
+}
+export {}
