+++ src/commands/db/diff.ts
@@ -1,0 +1,1 @@
+export {}
