+++ src/db-manifest.ts
@@ -1,0 +1,1 @@
+export {}
