+++ src/types/rules.d.ts
@@ -11,7 +11,4 @@ declare global {
     /** Function to get any data required by this rule */
     getData?: () => Record<string, unknown> | Promise<Record<string, unknown>>
   }
-
-  /** Allow TEMPLATE_DATA to be used as a global in rule definitions */
-  const TEMPLATE_DATA: Record<string, any>
 }
